refactor rest client
